fix : 자격증 서비스 수정, 강의노트 버그 수정 - 자격증 서비스 : 자격증 정보에 0회차 정보가 포함된 경우를 처리 - 강의노트 1. 유튜브 재생시간처리 부분 수정 2. 강이노트 추가 부분 수정

+++ src/test/java/com/steelrain/springboot/lilac/repository/LectureNoteRepositoryTest.java
@@ -13,7 +13,7 @@ class LectureNoteRepositoryTest {
 
     @Test
     void findTotalProgress() {
-        int tmp = m_lectureNoteRepository.findTotalProgress(2L, 9L, 23L);
+        long tmp = m_lectureNoteRepository.findTotalProgress(2L, 9L, 23L);
         assertThat(tmp >= 0).isTrue();
     }
 }
